Fix narrowing conversion warning (size_t -> unsigned int)

+++ src/game/map/Map.cpp
@@ -60,8 +60,8 @@ namespace game
 
         raylib::core::Vector2u Map::getPlayerStartingPosition(game::User::UserId playerId)
         {
-            return {(static_cast<size_t>(playerId) % 2) * (_size.x - 1),
-                (static_cast<size_t>(playerId) / 2) * (_size.y - 1)};
+            return {(static_cast<unsigned int>(playerId) % 2) * (_size.x - 1),
+                (static_cast<unsigned int>(playerId) / 2) * (_size.y - 1)};
         }
 
         void Map::freeCorners()
